Blocks now support IDs
